projects/sandbox/vizapp/pyproject.toml
[tool.poetry]
name = "vizapp"
version = "0.1.0"
description = "Visualization application for analyzed BBHNet output"
authors = ["Alec Gunny <alecg@.mit.edu>"]
license = "MIT"

[tool.poetry.scripts]
vizapp = "vizapp.main:main"

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
bokeh = "^2.4.3"
scipy = "^1.7"
h5py = "^3.5"

"hermes.typeo" = "^0.1.5"
"bbhnet.analysis" = {path = "../../../libs/analysis", develop = true}
"bbhnet.logging" = {path = "../../../libs/logging", develop = true}
"bbhnet.base" = {path = "../../../libs/base", develop = true, extras = ["numpy", "gwpy"]}
lalsuite = "^7.10"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
